+++ DownFall/Extensions/CGSize+Extension.swift
@@ -9,6 +9,10 @@
 import CoreGraphics
 
 extension CGSize {
+    
+    static var oneHundred = CGSize(width: 100, height: 100)
+    static var oneFifty = CGSize(width: 150, height: 150)
+    
     var playableRect: CGRect {
         /// This is a special ratio based on the iPhoneX ratio.  However we also account for the safe areas so we dont show anything too close to the top or bottom of the screen.  The side safe areas are ignored because we dont put anything useful in those areas to begin with
 //        let maxAspectRatio : CGFloat = 17.60/9.0
